Fix: Cross-Compile
A lot of automatic build tools, Travis CI, GitLab CI etc. are using docker containers
to build software. Unqlite can be cross-compiled for Windows from Linux.
This requires the Ming2-64 Cross Compiler.

However, within the source code there is an include for <Windows.h> this is correct for
Windows platforms but does not work when cross compiling.

In order to build a Windows binary of Unqlite from a Linux host, the include header has
to be changed to <windows.h>. In order to preserve cross-platform source code,
this commit fixes this issues to allow both succesfull compilation from Windows targets
as well as cross-compiling from a Linux host.

This fix allows Unqlite to be build by CI environments.